automatic_prompt/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "automatic-prompt-optimizer"
version = "0.1.0"
description = "Automatic Prompt Optimization (APO) - 基于论文实现的提示词自动优化系统"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Lu Ming", email = "your.email@example.com"}
]
keywords = ["prompt", "optimization", "llm", "ai", "nlp", "qwen"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "numpy>=1.21.0",
    "matplotlib>=3.5.0",
    "dashscope>=1.14.0",
    "python-dotenv>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "isort>=5.12.0",
]
openai = [
    "openai>=1.0.0",
]
anthropic = [
    "anthropic>=0.3.0",
]
local = [
    "transformers>=4.30.0",
    "torch>=2.0.0",
]
metrics = [
    "nltk>=3.8",
    "rouge-score>=0.1.2",
    "bert-score>=0.3.13",
]
all = [
    "automatic-prompt-optimizer[dev,openai,anthropic,local,metrics]",
]

[project.urls]
Homepage = "https://github.com/yourusername/automatic-prompt-optimizer"
Documentation = "https://github.com/yourusername/automatic-prompt-optimizer/blob/main/README.md"
Repository = "https://github.com/yourusername/automatic-prompt-optimizer"
"Bug Tracker" = "https://github.com/yourusername/automatic-prompt-optimizer/issues"

[tool.setuptools]
packages = ["automatic_prompt_optimizer"]

[tool.setuptools.package-data]
automatic_prompt_optimizer = ["*.md", "*.txt"]

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--verbose",
    "--cov=automatic_prompt_optimizer",
    "--cov-report=html",
    "--cov-report=term-missing",
]

[tool.coverage.run]
source = ["automatic_prompt_optimizer"]
omit = [
    "*/tests/*",
    "*/test_*.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
